change to command line interface using argparse and refactor to use a utility written by Brent Pedersen
Changed to a command-line interface using argparse,  and refactored to use a utility written by Brent Pedersen for the socket calls which is cleaner. Also included ability to get .bam and .bai files from CeNDR and a loop that ensures that IGV is running and tries to open it if it is not.